Reorganize folder structures in game
