+++ frontend/src/game/game_components/component.js
