+++ frontend/src/game/game_components/hitpoint.js
