+++ frontend/src/game/game_components/movement.js
@@ -1,7 +1,7 @@
 import Component from './component';
 import Input from './input';
 import Transform from './transform';
-import Time from './time';
+import Time from '../time';
 import Collider from './collider';
 import Vector from '../vector';
