+++ frontend/src/game/game_components/range.js
