+++ frontend/src/game/game_objects/create_bullet.js
@@ -1,9 +1,9 @@
 import GameObject from './game_object';
-import Transform from './transform';
-import BulletRenderer from './renderers/bullet_renderer';
-import Collider from './collider';
-import Circle from './circle';
-import BulletMovement from './bullet_movement';
+import Transform from '../game_components/transform';
+import BulletRenderer from '../renderers/bullet_renderer';
+import Collider from '../game_components/collider';
+import Circle from '../shapes/circle';
+import BulletMovement from '../game_components/bullet_movement';
 import Vector from '../vector';
 
 const createBullet = ({ id, position, directionVector, rotation, owned }) => {
