+++ frontend/src/game/game_objects/create_hitpoint_bar.js
@@ -1,16 +1,16 @@
 import GameObject from './game_object';
-import Transform from './transform';
-import TransformSyncronizer from './transform_syncronizer';
-import Syncronizer from './syncronizer';
-import CircleRenderer from './renderers/circle_renderer';
-import HitpointRenderer from './renderers/hitpoint_renderer';
-import Vector from '../vector'; 
-import Input from './input';
-import Movement from './movement';
-import Camera from './camera'; 
-import Shoot from './shoot';
-import Hitpoint from './hitpoint';
-import Circle from './circle';
+import Transform from '../game_components/transform';
+import TransformSyncronizer from '../transform_syncronizer';
+import Syncronizer from '../syncronizer';
+import CircleRenderer from '../renderers/circle_renderer';
+import HitpointRenderer from '../renderers/hitpoint_renderer';
+import Vector from '../vector';
+import Input from '../game_components/input';
+import Movement from '../game_components/movement';
+import Camera from '../game_components/camera';
+import Shoot from '../game_components/shoot';
+import Hitpoint from '../game_components/hitpoint';
+import Circle from '../shapes/circle';
 
 
 const createHitpointBar = ({ id, owned, position }) => {
@@ -22,13 +22,13 @@ const createHitpointBar = ({ id, owned, position }) => {
   player.addComponent(hitpoint);
   new Syncronizer(id+'1', hitpoint);
   player.addComponent(new CircleRenderer(radius, '#fce5cd'));
-  
+
   player.addComponent(new Movement());
   player.addComponent(new Collider(new Circle(radius)));
 
   if (owned) {
     player.addComponent(new Input());
-    player.addComponent(new Shoot()); 
+    player.addComponent(new Shoot());
     player.addComponent(new Camera());
   }
   return player;
