+++ frontend/src/game/game_objects/game_object.js
@@ -1,10 +1,10 @@
-import Game from './game';
+import Game from '../game';
 
 class GameObject {
   constructor(id, sort) {
     this.id = id;
     this.components = [];
-    this.sort = sort || 0; 
+    this.sort = sort || 0;
   }
 
   addComponent(component) {
@@ -13,13 +13,13 @@ class GameObject {
   }
 
   getComponent(componentClass) {
-    return this.components.find(component => 
+    return this.components.find(component =>
       component instanceof componentClass
     );
   }
 
   getComponents(componentClass) {
-    return this.components.filter(component => 
+    return this.components.filter(component =>
       component instanceof componentClass
     );
   }
