+++ frontend/src/game/shapes/rectangle.js
